chore: mocks generation for feature automation
Opted for a hierarchical structure of rake files.

With the idea of making the feature as independent as possible everyone
comes with its own rake file to build it self. A couple of paths are
still hardcoded but the idea is to inject those paths with rake in the
future.

closes #12