feat: logout and login with proper memory management of childs
Avoid having direct instances of chind coordinators and instead using
the collection provided by the coordinator protocol.

checked that the memory is freed up properly and all coordinators
stopped and freed.

I noticed and issue with the Collection coordinator but it is not due to
this job but to the Interactor-View-Presenter set up on that module. I
created a bug for that.

Closes #16