feat: Connect the MCS to the LoggedIn Coordinator
Create the bridge (`URLConsistencyHandlerBridge`) that connects the MCS
and the Lecture repository avoiding the memory loop.
Create a repository factory offered from the logged in coordinator to
the rest of the coordinators and dapters allowing them to create
repositories that keeps consistency.

WARNING: There is a knows issue of not persisting the data and also some
test that might be broken. That'll be fixed in the next commit.

Issue #23