refactor: Manage partial URL instead of the full URL
Before we were using the full URL everywhere and that was causing issues
due to the change of documents paths on every execution on the
simulator.

To avoid this and other issues in the future we are now deaing with a
partial URL with a URLComponents, and saving it as a String on the DB.
The full url is only obtained when needed on the LectureEntities.

Notice that the File System still uses internally the URL to refer to
the real files in the file system.

Closes #23